Module Std.Dwarf

Dwarf library This library gives an access to debugging information stored in a binary program.

module Section : sig ... end

File sections

module Tag : sig ... end

Debug Entry Tag

module Attr : sig ... end

Attribute

type lenspec =

| Leb128

| One

| Two

| Four

| Eight

val sexp_of_lenspec : lenspec -> Ppx_sexp_conv_lib.Sexp.t

val lenspec_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> lenspec

val bin_shape_lenspec : Bin_prot.Shape.t

val bin_size_lenspec : lenspec Bin_prot.Size.sizer

val bin_write_lenspec : lenspec Bin_prot.Write.writer

val bin_writer_lenspec : lenspec Bin_prot.Type_class.writer

val bin_read_lenspec : lenspec Bin_prot.Read.reader

val __bin_read_lenspec__ : (int -> lenspec) Bin_prot.Read.reader

val bin_reader_lenspec : lenspec Bin_prot.Type_class.reader

val bin_lenspec : lenspec Bin_prot.Type_class.t

val compare_lenspec : lenspec -> lenspec -> int

module Form : sig ... end

Attribute form

type tag = Tag.t

val sexp_of_tag : tag -> Ppx_sexp_conv_lib.Sexp.t

val tag_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> tag

val bin_shape_tag : Bin_prot.Shape.t

val bin_size_tag : tag Bin_prot.Size.sizer

val bin_write_tag : tag Bin_prot.Write.writer

val bin_writer_tag : tag Bin_prot.Type_class.writer

val bin_read_tag : tag Bin_prot.Read.reader

val __bin_read_tag__ : (int -> tag) Bin_prot.Read.reader

val bin_reader_tag : tag Bin_prot.Type_class.reader

val bin_tag : tag Bin_prot.Type_class.t

val compare_tag : tag -> tag -> int

type attr = Attr.t

val sexp_of_attr : attr -> Ppx_sexp_conv_lib.Sexp.t

val attr_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> attr

val bin_shape_attr : Bin_prot.Shape.t

val bin_size_attr : attr Bin_prot.Size.sizer

val bin_write_attr : attr Bin_prot.Write.writer

val bin_writer_attr : attr Bin_prot.Type_class.writer

val bin_read_attr : attr Bin_prot.Read.reader

val __bin_read_attr__ : (int -> attr) Bin_prot.Read.reader

val bin_reader_attr : attr Bin_prot.Type_class.reader

val bin_attr : attr Bin_prot.Type_class.t

val compare_attr : attr -> attr -> int

type form = Form.t

val sexp_of_form : form -> Ppx_sexp_conv_lib.Sexp.t

val form_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> form

val bin_shape_form : Bin_prot.Shape.t

val bin_size_form : form Bin_prot.Size.sizer

val bin_write_form : form Bin_prot.Write.writer

val bin_writer_form : form Bin_prot.Type_class.writer

val bin_read_form : form Bin_prot.Read.reader

val __bin_read_form__ : (int -> form) Bin_prot.Read.reader

val bin_reader_form : form Bin_prot.Type_class.reader

val bin_form : form Bin_prot.Type_class.t

val compare_form : form -> form -> int

type section = Section.t

val sexp_of_section : section -> Ppx_sexp_conv_lib.Sexp.t

val section_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> section

val bin_shape_section : Bin_prot.Shape.t

val bin_size_section : section Bin_prot.Size.sizer

val bin_write_section : section Bin_prot.Write.writer

val bin_writer_section : section Bin_prot.Type_class.writer

val bin_read_section : section Bin_prot.Read.reader

val __bin_read_section__ : (int -> section) Bin_prot.Read.reader

val bin_reader_section : section Bin_prot.Type_class.reader

val bin_section : section Bin_prot.Type_class.t

val compare_section : section -> section -> int

type fn

val bin_shape_fn : Bin_prot.Shape.t

val bin_size_fn : fn Bin_prot.Size.sizer

val bin_write_fn : fn Bin_prot.Write.writer

val bin_writer_fn : fn Bin_prot.Type_class.writer

val bin_read_fn : fn Bin_prot.Read.reader

val __bin_read_fn__ : (int -> fn) Bin_prot.Read.reader

val bin_reader_fn : fn Bin_prot.Type_class.reader

val bin_fn : fn Bin_prot.Type_class.t

val compare_fn : fn -> fn -> int

val sexp_of_fn : fn -> Ppx_sexp_conv_lib.Sexp.t

val fn_of_sexp : Ppx_sexp_conv_lib.Sexp.t -> fn

module Fn : sig ... end

Function representation.

module Buffer : sig ... end

Buffer is a light abstraction over string and bigstring, that can allow one to share the same string for different sections without explicit copying.

module Data : sig ... end

module Fbi : sig ... end

Function boundary identification.