Module Dwarf.Fbi

Function boundary identification.

type t

val create : string Data.t -> t Core_kernel.Or_error.t

create data tries to create a DWARF reader, from supplied data. May yield an error, if there wasn't sufficient sections, or if format is not understandable.

To provide information about functions parser needs at least this three sections:

.debug_abbrev Section.Abbr

.debug_info Section.Info

.debug_str Section.Str

val functions : t -> (string * fn) Bap.Std.seq

functions searcher enumerates functions